refactor(sync): slim the coordinator and gate syncs to resumed-only
- Drop the unused state-broadcast surface (SyncCoordinatorState class,
  StreamController, stream/state getters, _emit) — it had zero
  consumers;
  relocate the still-thrown SyncCoordinatorException into the
  coordinator.
- Collapse the ~15 fine-grained log lines into one sanitized summary per
  sync() (info on success, severe with kind:runtimeType on failure so
  Sentry
  still surfaces it).
- Gate syncs to AppLifecycleState.resumed only (was:
  not-paused-and-not-hidden,
  which also let inactive/detached through). On return to resumed,
  replay a
  catch-up sync so a request gated while away is never lost.
